+++ packages/editor/editor-common/src/utils/calc-breakout-width.ts
@@ -18,7 +18,7 @@ export const calcBreakoutWidth = (
     case 'wide':
       return calcWideWidth(containerWidth);
     default:
-      return 'inherit';
+      return '100%';
   }
 };