+++ packages/media/media-viewer/examples/0-single-file-previews.tsx
@@ -26,8 +26,8 @@ import { I18NWrapper } from '@atlaskit/media-test-helpers';
 const context = createStorybookContext();
 
 const handleEvent = (analyticsEvent: UIAnalyticsEventInterface) => {
-  const { payload, context } = analyticsEvent;
-  console.log('Received event:', { payload, context });
+  const { payload } = analyticsEvent;
+  console.log('EVENT:', payload);
 };
 
 export type State = {