+++ packages/media/media-image/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-image
 
+## 7.0.9
+- [patch] [ef469cbb0b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ef469cbb0b):
+
+  - MS-357 replaced @atlaskit/util-shared-styles from media components by @atlaskit/theme
+
 ## 7.0.8
 - Updated dependencies [69c8d0c19c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/69c8d0c19c):
   - @atlaskit/media-test-helpers@20.1.0