+++ packages/media/media-ui/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-ui
 
+## 8.5.1
+- [patch] [ef469cbb0b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ef469cbb0b):
+
+  - MS-357 replaced @atlaskit/util-shared-styles from media components by @atlaskit/theme
+
 ## 8.5.0
 - [minor] [be66d1da3a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/be66d1da3a):